src/requirements.txt
openai==0.27.8
python-dotenv==1.0.0
langchain==0.0.242
beautifulsoup4==4.12.2
faiss-cpu==1.7.4
tiktoken==0.4.0
PyQt5